refactor(pdf): isolate native vector artifact classification
